requirements.txt
asgiref==3.3.4
attrs==20.3.0
coverage==5.5
Django==3.2
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.1.0
flake8==3.9.0
iniconfig==1.1.1
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
